requirements.txt
# Веб фреймворки
fastapi
uvicorn[standard]
pydantic-settings

# Брокер сообщений 
faststream[rabbit,cli]

# БД и ORM
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
alembic

# Скрапинг и Утилиты
curl_cffi
httpx
beautifulsoup4
aiofiles
playwright>=1.40.0

redis 
qdrant-client

# Тесты
pytest
pytest-cov